[#52] removed extra logging for nonexistant files.
